+++ src/main.cpp
@@ -1,13 +1,13 @@
 
-#include "modules/gui.h"
-#include "modules/input.h"
-#include "modules/main_menu.h"
-#include "modules/phase.h"
-#include "modules/render.h"
-#include "modules/rocket_launch.h"
-#include "modules/simulation.h"
-#include "modules/site.h"
-#include "modules/staff.h"
+#include "modules/gui/gui.h"
+#include "modules/input/input.h"
+#include "modules/main_menu/main_menu.h"
+#include "modules/phase/phase.h"
+#include "modules/render/render.h"
+#include "modules/rocket_launch/rocket_launch.h"
+#include "modules/simulation/simulation.h"
+#include "modules/site/site.h"
+#include "modules/staff/staff.h"
 #include "spdlog/sinks/basic_file_sink.h"
 #include "spdlog/spdlog.h"
