+++ src/modules/gui/gui.cpp
@@ -2,8 +2,8 @@
 #include "backends/imgui_impl_sdl2.h"
 #include "backends/imgui_impl_sdlrenderer2.h"
 #include "imgui.h"
-#include "modules/phase.h"
-#include "modules/render.h"
+#include "modules/phase/phase.h"
+#include "modules/render/render.h"
 #include "spdlog/spdlog.h"
 #include <flecs.h>
