+++ src/modules/site/site.cpp
@@ -1,11 +1,10 @@
 #include "site.h"
 #include "SDL_keycode.h"
-#include "flecs/addons/cpp/c_types.hpp"
 #include "imgui.h"
-#include "modules/input.h"
-#include "modules/phase.h"
-#include "modules/rocket_launch.h"
-#include "modules/simulation.h"
+#include "modules/input/input.h"
+#include "modules/phase/phase.h"
+#include "modules/rocket_launch/rocket_launch.h"
+#include "modules/simulation/simulation.h"
 #include "spdlog/spdlog.h"
 #include "widgets/widgets.h"
